requirements.txt
attrs==20.3.0
autopep8==1.5.6
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
google-api-core==1.26.1
google-auth==1.28.0
google-cloud-language==2.0.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
idna==2.10
importlib-metadata==3.7.3
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
libcst==0.3.17
MarkupSafe==1.1.1
mypy-extensions==0.4.3
newsapi-python==0.2.6
packaging==20.9
pluggy==0.13.1
proto-plus==1.18.0
protobuf==3.15.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyparsing==2.4.7
pytest==6.2.2
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.4
Werkzeug==1.0.1
zipp==3.4.1
